refactor register for multiple commands
