+++ src/commands/commands.ts
@@ -1,0 +1,7 @@
+import bookclubCommands from "./bookclub/bookclub";
+
+const commands = {
+  bookclubCommands,
+};
+
+export default commands;
