+++ src/commands/register.ts
@@ -1,47 +1,24 @@
-import { SlashCommandBuilder } from "@discordjs/builders";
 import { REST } from "@discordjs/rest";
 import { Routes } from "discord-api-types/v9";
+const { bookclubCommands } = require("./commands");
 require("dotenv").config();
 
 const BC_TOKEN = process.env.BOOK_CLUB_BOT_TOKEN_ID;
+const GUILD_ID = process.env.GUILD_ID;
 
-const commands = [
-  new SlashCommandBuilder()
-    .setName("bookclub")
-    .setDescription("Interact with the Space Book Club")
-    .addSubcommandGroup((group) =>
-      group
-        .setName("recommend")
-        .setDescription("Get a book recommendation")
-        .addSubcommand((option) =>
-          option
-            .setName("random")
-            .setDescription("Get a random book from the app")
-        )
-        .addSubcommand((option) =>
-          option
-            .setName("best")
-            .setDescription("Get the highest ranked book from the app")
-        )
-        .addSubcommand((option) =>
-          option
-            .setName("favourite")
-            .setDescription("Get the community's favourite book from the app")
-        )
-    )
-    .addSubcommand((option) =>
-      option.setName("help").setDescription("Get help with the Book Club bot")
-    ),
-].map((command) => command.toJSON());
+const register = (token) => {
+  const rest = new REST({ version: "9" }).setToken(token);
 
-const rest = new REST({ version: "9" }).setToken(BC_TOKEN);
+  return rest.put(
+    Routes.applicationGuildCommands("781234878992744488", GUILD_ID),
+    {
+      body: bookclubCommands,
+    }
+  );
+};
 
-console.log(BC_TOKEN);
-
-rest
-  .put(
-    Routes.applicationGuildCommands("781234878992744488", "781235493118672946"),
-    { body: commands }
-  )
-  .then(() => console.log("Successfully registered application commands."))
+Promise.all([register(BC_TOKEN)])
+  .then(() => {
+    console.log("Successfully registered application commands.");
+  })
   .catch(console.error);
